Agent: Move agent_event_forwarder.py to agent_event_handlers package
