+++ monkey/infection_monkey/agent_event_handlers/__init__.py
@@ -1,1 +1,2 @@
 from .notify_relay_on_propagation import notify_relay_on_propagation
+from .agent_event_forwarder import AgentEventForwarder
